Moved old main to documents, loluino loads main
