+++ platformio.ini
@@ -18,7 +18,7 @@ board_build.f_cpu = 240000000L
 framework = arduino
 lib_deps = Memory_Manager_lib_d, simple_lib_d
 ; new flags start after gc--sections
-build_flags =-fpic, -Wl,--gc-sections, -Wl,-static -fpermissive
+build_flags =-fpic, -Wl,--gc-sections, -Wl,-static -fpermissive 
 ;-D CONFIG_SPIRAM_ALLOW_BSS_SEG_EXTERNAL_MEMORY, -DMALLOC_CAP_SPIRAM
 ; s removes debug table, 
 board_build.f_flash = 40000000L ;https://community.platformio.org/t/a-question-about-board-configuration-files-with-esp32/21877/6 Added config stuff for flash from here
@@ -27,7 +27,8 @@ extra_scripts = post:extra_script.py ;
 
 build_unflags= "-g3 --no-text-section-literals"
 test_ignore = test_desktop
-
+;test_speed =460800 ; this is new. It was copied from what build changes the rate to when used with this board.
+;build_flags =verbose
 [env:firebeetle32]
 platform = espressif32
 board = firebeetle32
